before adding smart ptrs

+++ fall17-proj1source/Puzzle8State.h
@@ -173,7 +173,7 @@ public:
 		return fValue;
 	}
 
-	void SetFValue(int f) {
+	void SetFValue(double f) {
 		fValue = f;
 	}
